% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurateappendr.R
\docType{package}
\name{accurateappendr}
\alias{accurateappendr}
\alias{accurateappendr-package}
\title{accurateappendr: R Interface to AccurateAppend API}
\description{
AccurateAppend is a marketing service to help fill in missing contact info
for people in your contact list.
}

