% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{consumer_phone}
\alias{consumer_phone}
\title{Search For a Single Consumer's Phone Numbers}
\usage{
consumer_phone(key, first_name = NA, last_name, address = NA, city, state,
  postal_code = NA, sources = c("C", "D", "M", "E"),
  match_levels = c("E1", "E2", "N1", "N2", "B1", "B2"))
}
\arguments{
\item{key}{Your AccurateAppend API Key}

\item{first_name}{String}

\item{last_name}{String, required}

\item{address}{String}

\item{city}{String, required}

\item{state}{String, state postal abbreviation, required}

\item{postal_code}{String}
}
\value{
Data frame with 6 columns:
\describe{
  \item{AreaCode}{Character}
  \item{LineType}{Character, mobile v. landline}
  \item{MatchLevel}{Character, code for fidelity of match}
  \item{MaxValidationLevel}{Character, maximum MatchLevel possible for data provided}
  \item{PhoneNumber}{Character}
  \item{Source}{Character, code for source database}
}

See \url{http://docs.accurateappend.com/docs/appendphone} for additional
infor including match code and source definitions.
}
\description{
Search For a Single Consumer's Phone Numbers
}

